security Changes

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,16 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index','show']]);
+    }
+    
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -75,7 +85,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post= Post::find($id);
-        return view('posts/edit')->with('post',$post);
+        //Check users Id
+        if(auth()->user()->id == $post->user_id ){
+            return view('posts/edit')->with('post',$post);
+        }
+        return redirect('/posts')->with('error','UnAuthorized Page!');
     }
 
     /**
@@ -104,7 +118,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
-        $post->delete();
-        return redirect('/posts')->with('success','Post Deleted Successfully');
+        //Check users Id
+        if(auth()->user()->id == $post->user_id ){
+            $post->delete();
+            return redirect('/posts')->with('success','Post Deleted Successfully');
+        }
+        return redirect('/posts')->with('error','UnAuthorized Page!');
     }
 }
